Split common.go into parts for all OSes and Linux-only parts
